+++ packages/tailwindcss-patch/src/cli.ts
@@ -1,4 +1,4 @@
-import { createPatch, getPatchOptions } from './patcher'
+import { createPatch, getPatchOptions } from './core/patcher'
 const opt = getPatchOptions()
 const patch = createPatch(opt)
 patch()
