+++ packages/tailwindcss-patch/src/core/index.ts
@@ -1,0 +1,4 @@
+export * from './class'
+export * from './exposeContext'
+export * from './inspector'
+export * from './patcher'
