+++ packages/tailwindcss-patch/src/index.ts
@@ -1,6 +1,3 @@
-export * from './class'
-export * from './exposeContext'
-export * from './inspector'
-export * from './patcher'
+export * from './core'
 export * from './utils'
-export * from './type'
+export * from './types'
